requirements.txt
streamlit
pandas
numpy
requests
python-deriv-api